Feat: add integration tests and Unit tests

+++ db/migrate/20231010170055_create_recipes.rb
@@ -2,8 +2,8 @@ class CreateRecipes < ActiveRecord::Migration[7.0]
   def change
     create_table :recipes do |t|
       t.string :name
-      t.integer :preparation_time
-      t.integer :cooking_time
+      t.float :preparation_time
+      t.float :cooking_time
       t.text :description
       t.boolean :public, default: true
       t.references :user, null: false, foreign_key: true
